frontend/requirements.txt
# Frontend dependencies

# Streamlit and UI
streamlit==1.28.0
streamlit-chat==0.1.1
streamlit-mic-recorder==0.0.4
streamlit-option-menu==0.3.6
streamlit-feedback==0.1.3
streamlit-pdf-viewer==0.0.2

# Document processing
pypdf2==3.0.1
python-docx==1.0.1
docx2pdf==0.1.8
reportlab==4.0.4

# Speech processing
speechrecognition==3.10.0
gTTS==2.3.2
pydub==0.25.1

# HTTP requests
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pillow==10.0.1
numpy==1.26.0
pandas==2.1.1